+++ contrib/etc/_appsettings.Development.json
@@ -2,7 +2,7 @@
   "Jwt": {
     "Issuer": "FunderMapsDev",
     "Audience": "FunderMapsDev",
-    "TokenValidity": 5,
+    "TokenValidity": 60,
     "SignatureKey": "dhfgdfubhgbdfghredfgregurhe78gh348yvt784tcno3qt74gntkuirdhg"
   },
   "Logging": {
