+++ tests/FunderMaps.IntegrationTests/Backend/Application/AuthTests.cs
@@ -1,6 +1,5 @@
 ﻿using Bogus;
 using FunderMaps.AspNetCore.DataTransferObjects;
-using FunderMaps.AspNetCore.InputModels;
 using FunderMaps.Testing.Extensions;
 using System;
 using System.Net;
@@ -54,7 +53,7 @@ namespace FunderMaps.IntegrationTests.Backend.Application
             using var client = Factory.CreateUnauthorizedClient();
 
             // Act
-            var response = await client.PostAsJsonAsync("api/auth/signin", new SignInInputModel()
+            var response = await client.PostAsJsonAsync("api/auth/signin", new SignInDto()
             {
                 Email = "lester@contoso.com",
                 Password = new Randomizer().Password(64),
@@ -78,7 +77,7 @@ namespace FunderMaps.IntegrationTests.Backend.Application
             using var client = Factory.CreateUnauthorizedClient();
 
             // Act
-            var response = await client.PostAsJsonAsync("api/auth/signin", new SignInInputModel()
+            var response = await client.PostAsJsonAsync("api/auth/signin", new SignInDto()
             {
                 Email = email,
                 Password = password,
@@ -96,7 +95,7 @@ namespace FunderMaps.IntegrationTests.Backend.Application
             using var client = Factory.CreateUnauthorizedClient();
 
             // Act
-            var response = await client.PostAsJsonAsync("api/auth/signin", new SignInInputModel()
+            var response = await client.PostAsJsonAsync("api/auth/signin", new SignInDto()
             {
                 Email = email,
                 Password = password,
